requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
joblib==1.2.0
numpy==1.24.4
pandas==2.2.3
scikit-learn==1.2.2
scipy==1.10.1
supabase==2.11.0
python-dotenv==1.0.1
